requirements.txt
Django==5.1.6
djangorestframework==3.15.2
openpyxl==3.1.5
python-decouple==3.8
